Foto, Plato, mesa, conf seg

+++ web/src/main/java/com/RestoApp2/web/Entidades/Plato.java
@@ -1,8 +1,4 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
+
 package com.RestoApp2.web.Entidades;
 
 import com.RestoApp2.web.Enums.Categoria;
@@ -13,10 +9,6 @@ import javax.persistence.Id;
 import javax.persistence.OneToOne;
 import org.hibernate.annotations.GenericGenerator;
 
-/**
- *
- * @author Federico
- */
 @Entity
 public class Plato implements Serializable{
     
@@ -27,7 +19,7 @@ public class Plato implements Serializable{
     
     private String nombre;
     private String descri;
-    private Integer precio;
+    private Double precio;
     @OneToOne
     private Foto foto;
     private Boolean alta;
@@ -36,7 +28,7 @@ public class Plato implements Serializable{
     public Plato() {
     }
 
-    public Plato( String nombre, String descri, Integer precio, Foto foto, Boolean alta, Categoria categoria) {
+    public Plato( String nombre, String descri, Double precio, Foto foto, Boolean alta, Categoria categoria) {
         
         this.nombre = nombre;
         this.descri = descri;
@@ -70,11 +62,11 @@ public class Plato implements Serializable{
         this.descri = descri;
     }
 
-    public Integer getPrecio() {
+    public Double getPrecio() {
         return precio;
     }
 
-    public void setPrecio(Integer precio) {
+    public void setPrecio(Double precio) {
         this.precio = precio;
     }
